Fix to charge-self-consistency with Vasp and QE
PR #48 and #46

The occupation correction GAMMA in Vasp is now recalculated between subsequent steps. Also reorganized csc_flow to reflect the actual process better and make it more readable. Other changes:

- Introduced parameter dft/n_iter_first: the number of DFT iterations in the first round of charge corrections
- Removed parameter general/oneshot_postproc_gamma_file
- Added docker file with Vasp 5 to avoid problems with the current Vasp 6 setup for CSC
- Moved the more technical functions from csc_flow into the dft_managers
- Changed the way QE restarts a calculation because before it ended up diverging frequently when restarting
- Made w90 the default for parameter dft/projector_type: this option is actually supported by both DFT codes
- solid_dmft raises an error when trying to restart Vasp CSC because it's not cleanly implemented

These changes are tested with Vasp on SrVO3 with Wannier90 and PLO projectors and on LuNiO3 with PLO projectors as well as in QE with Wannier90 on LuNiO3.

Also fixes total energy with QE as described in PR #46